Add command line mode for NetworkManager module
